Basic Parsing for Develop Settings
Basic work in place to parse the settings - parser method, test class and data class for the settings I intend to extract. There is quite a bit of interesting data about edits in the settings and of note to even basic display of the data is (sometimes anyway) better than EXIF lens information.
Db connection string to ReadOnly.